MET-1111 Changed the data type of sigma to double
Former-commit-id: 192d3cea4b2069edf78328cfc1835e8d3fae13c9

+++ met/src/libcode/vx_statistics/pair_base.cc
@@ -783,7 +783,7 @@ double compute_interp(const DataPlaneArray &dpa,
                       const double thresh,
                       const bool spfh_flag, const LevelType lvl_typ,
                       const double to_lvl, const int i_blw, const int i_abv,
-                      const SingleThresh *cat_thresh, const int sigma) {
+                      const SingleThresh *cat_thresh, const double sigma) {
    double v, v_blw, v_abv, t;
 
    // Check for no data
